+++ pages/_app.tsx
@@ -3,7 +3,7 @@ import { ThemeProvider } from 'theme-ui';
 import { AppProps } from 'next/app';
 import theme from '@cravend/theme';
 import { init } from '@socialgouv/matomo-next';
-import { withAutomaticColorMode } from '../src/useColorMode';
+import { withAutomaticColorMode } from '../src/hooks/useColorMode';
 
 const App:FC<AppProps> = ({ Component, pageProps }: AppProps) => {
   useEffect(() => {
